feat：添加了路径常量，添加了远程配置文件
