+++ syncl2r/config/constant.py
@@ -1,0 +1,8 @@
+from pathlib import PurePath
+
+SyncL2R_l2r_Path = PurePath("./.l2r")
+Data_Path = SyncL2R_l2r_Path / "data.yaml"
+History_Log_Path = SyncL2R_l2r_Path / "logs"
+Config_Path = SyncL2R_l2r_Path / "config.l2r.yaml"
+Temp_Output_Path = SyncL2R_l2r_Path / "out.log"
+Temp_Pids_Path = SyncL2R_l2r_Path / "pids.txt"
