+++ syncl2r/syncl2r.py
@@ -1,7 +1,7 @@
 from syncl2r.config import load_config
 
 from .command import app
-
+from .console import pprint
 
 __all__ = ["main"]
 
@@ -9,8 +9,8 @@ __all__ = ["main"]
 def main():
     try:
         load_config()
-    except:
-        pass
+    except Exception as e:
+        pprint(f"Some error happened, detail: {e}")
     app()
